Rewrite message rendering, and auto-scale images.
Rendering is now done entirely after the message has been
turned into a DOM object. This saves costly conversions
to HTML code and back, which also strip the event listeners.

This allows us to automatically scale images on loading,
as well as rescaling them when a menu is toggled to change the
chatList width.