Move formatTime and messageCreate to visual.js
It makes sense to have all the rendering code
in one place. This did require removing the message.id
counter, but we were not using that anyway:
The DOM references are stored in the message array when
added, so we don't need the unique ID attribute.

Also, renamed renderText to formatBody,
and messageCreate to formatMessage for consistency.