how does boolean?!
This line in insertText() is supposed to determine
whether to put the cursor at the end of the start tag
or the end of the replace text.
Obviously the former is only desired when no text is
selected (since the tag is already filled then), and
when actual BBCode is inserted.