xmpp bugfix: nick conflict rejoin didn't work.
xmpp.currentRoom is set only when the room is
successfully joined. However, we still tried
to rejoin xmpp.currentRoom (undefined) on a nick
conflict.

Instead, we now have xmpp.room.target (a la
xmpp.preferredNick) which is set to the room
we *want* to join.

Variables have been renamed for consistency.
Also, the room to join on connection has been
moved to config.settings to make it modifiable
(but isn't yet saved).