Add default /say command, escape with //.
We need a way to post messages with leading /
(other than /me), in chat; including commands
that are otherwise interpreted. This is now
done by escaping them with /say or with an
extra slash.

(This makes everything entered a command, and
thus prevents messages from being sent to the
server while disconnected.)