Add persistent color setting.
This setting only has an effect for other
cadence users. The rationale is that persistent
font colors - chiefly used to distinguish users,
and for stylistic expression need to be easily
disabled without interfering with other, more
important markup. XMPP clients do not generally
possess that level of fine-tuning.

Therefore, the persistent font color will be added
as a class attribute on the outgoing message.
cadence will then look for this class on incoming
messages and turn it into a css-rule.