Added /away and /back commands.
xmpp now has a sendStatus() function that will
send a presence to the current roomnick with
an optional <show> and <status> element.

Also, the status on incoming presences will
be shown in the notification if present; the
user-link will also be updated to display an
away icon when appropriate.

Some bugs in other functions were exposed and
fixed due to this feature.