Several output security fixes.
The greatest security consideration is script safety
of the recipient. Incoming Javascript could significantly
compromise the client, including the leakage of XMPP
session or authentication data.

We can't trust XMPP servers to keep HTML safe.
This filtering must be done by the recipient.
Therefore we pick a small whitelist (a,img,span,q,
code,strong,em,blockquote) of tags, and replace all
the others with their escaped outerHTML.

Furthermore, add a configurable limit (24/1024)
on nick and message length, which will be applied both
to incoming and outgoing messages.

Outgoing HTML will furthermore be escaped before
rendering BBCode. That is not real security, because
it happens on the client, but it will prevent
unintentional errors.