Rewrite formatMessage() with better /me handling.
Fixes #3: /me commands in XHTML are improperly parsed.

The pattern I have decided to go with instead is as follows:

* A me-node is a Text node whose first four
characters are '/me '.
* A me-message is a me-node, or has a me-message
  as its first child.
* The me-child of a me-message is itself if it is a
  me-node; otherwise it is the me-child of its first child.

Iff a message is a me-message, then it will be italicized,
the author prepended with '* ' (instead of appended with ':'),
and its me-child will have its first four characters truncated.

(Notably, this means /me can be contained inside any markup,
provided it is not preceded by any other element or text.
The markup around /me will not be applied to the author.)