Implement /msg command (fixes #2).
This patch adds the /msg command. It will
send a private message to the specified user,
provided that user is in the room.

As the server does not acknowledge a successful
private message, let alone reflect it, we print
out the message manually.

But that means it will be printed before any errors
arrive, so we check whether the recipient is valid
before trying to send.