Bugfix in join/leave flow.
On successfully joining a room, the current room (if present)
will be left and *then* room.current will be changed.

But the leaveRoom() function will put us in prejoin mode when
leaving a room that matches room.current. So it needs to be
executed *after* room.current has been changed.