Rewrite formatText with formatting types.
All variables passed to formatText now have a type, which is
either their name or (for multiple variables of the same type)
a dotted prefix. A variable named {type} or {type.x} will be
passed through visual.format.type() before output.

The default type is "plain", which escapes all markup.

Rendered variables are typed as "raw", which will bypass the
filter and print them directly.