Add /save command (fixes #11).
Thanks to yet another JS library, saving the log
as a file is now a thing.

The log can be saved as plain text (stripping all text,
but reverting links and images to BBcode to avoid losing
information, and reverting emoticons to text).
The log can also be saved as unchanged HTML, but currently
without any styling information and keeping emoticons as relative
links (which don't work).