Fix #42: Show delayed messages from offline users.
When a timestamped message arrives from a user who is not in the
online roster, create a fake user object from their nick, so it
can be displayed.

This means admins will mysteriously use their color in backlog
messages when they're logged out, but that cannot be helped: XMPP
does not specify user roles when transmitting these messages.