Add "/msg <nick>" when clicking on users (#5).
When clicking on a nickname anywhere, a /msg <nick>
is now prepended to the chat box.

This is only a temporary patch, and will be replaced by something cleaner
in jquery at some point. For now, the command is embedded in the markup
string, which requires some really awful quote-escaping code.