Fix #32: Fix custom styles in webkit.
The issue, as near as I can tell, is that webkit doesn't actually
apply an alt stylesheet when it is added to DOM, but requires the
disabled attribute to be jiggled.

Therefore, ui.setStyle now disables all stylesheets before enabling
one. Strictly speaking, it only needs to do this once on the initial
load, but it is trivial and probably more robust to do it every time.

This bug affected Google chrome 32.0.1700.77 and at least one
version of Safari, but not Chromium 31.0.1650.63.