Fix #25: Apply nick change to xmpp.nick.current
When a user is logged in from multiple resources as
the same roomnick, there is no 303 status when changing nicks.

(Furthermore, ejabberd does not send a 110 status on the
reflected presence; see issue 136 there.)

Therefore, work around this by detecting a nick change confirmation
based on the fact that the nick we wanted to change to has now entered
the room. The possibility that another user managed to change to that
nick at *just* the right time before we get a nick collision should be
basically negligible. (But hypothetically hilarious.)