Fix #67: Ignore self-nick until room is joined.
As of f67d9b4, for #25, the presence callback looks for a self-presence
that matches nick.target but not nick.current in order to confirm
a successful nick change request.

There is another circumstance that causes this: Joining a room when
nick.current is not set, and the same user is already in the room
with the same nickname, and listed in the roster.

Therefore, the check from f67d9b4 should be moved into the
room == this.room.current block, which is only executed after the roster
is complete.