Fix #76: Remove splash page; make alt styles static.
This patch moves the generation of alt stylesheet links from the
ui initializer to the static build process (on the reasonable assumption
that the application will be rebuilt when a new style is added).

This dramatically improves page load performance, because the stylesheets
will be loaded immediately, not after they have been added by Javascript.

The black splash page is now only visible if JS is disabled, or for
an instant while the CSS styles are loading. That makes the custom
logo superfluous, so that's gone too. The splash page is now not toggled
by Javascript, but by the alt stylesheets themselves.

Also fixes #56.