Fix #81 (I hope): replace rsync with cp.
The cp command is wrapped in a macro, so it can be
overridden by setting the CP environment variable.

Also tearing out the "core" target, which copies JS core sources
to the build folder. It's not referenced anymore, and in any case
I think it's obsolete since both aggregation and installation go
for the source files directly.