Fix #92: Don't copy files over themselves.
Installing to the source tree is explicitly supported.

In these cases, "make install" is not supposed to copy index.html;
without a CDN, it should not copy anything at all. Otherwise, cp
will return errors because the origin and destination are identical.
In that case, "install" should merely be an alias to "all".

Add a check for CDN_PREFIX and PREFIX, and make the install target behave
accordingly.