Fix #96 (for now): Show unknown users' messages.
There is some bug that occasionally causes users to not be in the roster
despite being logged in. Either ejabberd is not sending the presence,
or cadence is not processing it right.

To alleviate the worst effects of this bug, instead of silently dropping
messages from these invisible users, cadence will now handle them like
backlog messages from offline users: User is formatted without affiliation,
role or jid; just the nickname.

(PS: A little logic clean-up in the affected code block.)