Fix #104: Print error for bad nickname.
When the server sends a <jid-malformed /> presence error,
it is presumably a nick error because we check rooms with <iq>
before attempting to join them.

So when such an error arrives, cancel the nick change and
any join attempt (in case the bad nick was set before joining)
and alert user to the invalid nickname.