Fix #106: Expose JID in CSS classes.
This is a bit complicated. We want to style stuff based on user
identity. This can be done in a limited way by selecting the
title attribute of the span.user element, but it can neither
affect messages (due to missing up-traversal) nor take advantage
of class-selector optimization.

So this patch adds three classes to every div.message and
span.user-roster element (the latter is a new container to allow
adding "user flair" icons without clashing with away icons):

* jid-node-<local>
* jid-domain-<domain>
* jid-resource-<resource>

Any spaces in these values are escaped to \dec character codes.