Fix #83: Implement server commands.
This patch implements server commands as defined by XEP-0133.

The syntax is /admin <cmd> <arg>.

Two server commands, announce and set-motd, are added for now.
Note that cadence implements the variant used by ejabberd,
where instead of a "motd" or "announcement" field, a "body" field is
submitted for both commands.