Fix #119: unset connection on disconnect event.
If the xmpp.connection object is unset on calling xmpp.disconnect(),
rather than on an incoming disconnect event, then the connection object
will linger after a network or server error.
On the next reconnect, this will cause cadence to try to disconnect the
broken connection, and result in an error.