Fix #126: Alert sound on watched user entrance.
When a user enters whose nick matches an alert keyword, the "mention" sound
should be played instead of the "enter" sound. This allows e.g. keeping the
chat in the background while waiting for a specific user to enter, instead
of only being alerted when that user sends a message.