Re #50: Disable persistent color retroactively.
When persistent colors are disabled or enabled, this now affects
all messages in the chat history up to this point.

This makes it possible to switch themes and retain readable chat histories.

(This is the easiest of the three "retroactivity" patches, because
adding and removing color doesn't alter the vertical size of the history.)