(#152, continued): Update all libraries.
 - buzz
 - filesaver
 - jquery-cookie
 - moment
 - strophe

Also, stop rebuilding strophe. The strophe.js file in the repository is up to
date, and the build process now requires nodejs, which would be inconvenient
to add to our dependencies for just that purpose.
In fact, this part of the Makefile was left over from the time when we used
strophe.min.js instead of strophe.js, so it has long been unnecessary.