Fix #144: User roster index is broken.
ui.userRemove() should splice a particular user out of ui.sortedNicks
with ui.sortedNicks.splice(i, 1).

Instead, it does ui.sortedNicks.splice(i, i).

Removing more users is generally harmless as it will simply mess up the ordering.
But if the first user logs out, their entry will remain in the index even though the
corresponding DOM element is deleted.
Anyone with a nick that precedes it alphabetically will attempt to attach to a missing
DOM element and fail to show up in the list until the next reconnect.