Fix #171: Ignore phantom departures.
Kicking or banning a user who is logged in with multiple resources
causes ejabberd to broadcast "unavailable" presences for all their
resources simultaneously.

Because multiple resources are not tracked by cadence (because ejabberd
does not otherwise broadcast individual resources leaving), the user
is removed as soon as the first "unavailable" presence is received.
Additional presences then cause an error.

Therefore, "unavailable" presences are now silently ignored if they
do not match an existing roster entry.