Fix #175: Unify URL fragment updates.
The URL fragment needs to always reflect the current room.
This means it must also be updated in response to external events -
failure to join a room, being kicked from it, etc.

Additionally, the /join command had a bug that failed to strip a space
preceding the room name in the URL fragment.