Fix #196: Rewrite emoticon and URL rendering.
Remove the flawed approach of injecting HTML code into
text node content and then evaluating the result.

Instead, tokenize the text node content and insert the
rendered images (as well as the text between them) into
the DOM before removing the old text node.

This removes jquery-replacetext and replaces it with
a custom library solely dedicated to replacing RegExp
matches with DOM nodes.