Fix #236: Remove fake user.nick values.
The user.nick property is now strictly limited to room occupants,
and not a stand-in for the displayed name.

The formatter uses user.nick or the bare JID part of user.jid;
any functions that create a user object should only set user.nick
to represent the nickname of a room occupant.

(The bot is a special case.)