Fix #244: Destroy rooms.
The new /destroy [--reason <...>] [--alternate <...>] [<room>]
command will order the server to destroy a room (by default, the current room).

If a room is destroyed by the server, then the user will now properly
be put into prejoin mode and informed of the destruction (along with the message
or alternate room, if they are transmitted).