Fix #255: Explicitly style <option> element.
Chromium does not use the <body> background for dropdowns,
unlike Firefox. Instead, the dropdown is white by default,
and a semi-transparent <select> background is added to that.

The desired composite value must be set on <option> elements
explicitly to get a consistent effect in both browsers.
(This only effects styles with a transparent <select> element.)

Styles affected: Plum, Selenium, dash, omg.