Fix #292 1/5: Rewrite target JID generation.
xmpp functions that take a target now expect an object
instead of a variety of different argument styles.

The object will be passed to xmpp.jid().
(xmpp.jid has been renamed to xmpp.currentJid.)
xmpp.jid() makes some intelligent guesses on what defaults
to use for the given arguments, and returns a JID.