[#168] Expose the language setting.
A language menu is added to the General settings tab.
All available locales are added to it as options.
Changing the language instantly replaces all tagged strings
in the user interface.

A new key is also added to the root level of each locale: "language".
This string is the localized name of the language, used as the option label.