Fix #309: Block /ping, /version to bare JID.
IQ stanzas sent to a bare JID are handled by the server, not the client,
which means that a /ping or /version sent to them is meaningless.
Instead of looking for a '@' (which in any case is not foolproof),
these commands should check the return value of getResourceFromJid(),
and treat any value that isn't a full JID as a room-nick.