Fix #319: Reformat dmsg desktop notifications.
Add a new message type, "direct", which is set on all non-MUC messages.
It's not groupchat, so the messageFormat will still make it a whisper.

ui.notifyDesktop(), however, will use the sender's JID as the title
(instead of the active room), and will not prepend the sender or "(whispers)"
to the message.