Fix #363: Add a JID class.
This defines a JID class that can be created from a node, domain
and resource string, or parsed from an existing string value.

JID is a subclass of String, allowing code to continue using its
instances as string values.

Many superfluous parsing calls will be removed in a follow-up patch.