Fix #351: Status context menu.
The status context menu replaces the logout button
and connection indicator.
The menu allows setting the status (available, away, xa, dnd),
as well as connecting or disconnecting.
This patch increases the tray icon size to 24x24, and will cause
some visual problems until the remaining icons are adjusted as well.