[#261] Fix sidebar and settings application.
This rewrites ui.toggleMenu to respond to config changes, rather
than perform them. Now it can be used to update the sidebar after
the settings are synchronized.

ui.loadSettings() calls a .change() event on all settings elements,
to update things like timestamp formats and text colors.

We also add a data-sidebar attribute to the toggle buttons, because
parsing it out of their ID attribute kind of smells.