Fix #270: Persistent nickname.
This will save the nickname in the settings, and reapply it
on startup. To minimize astonishment, the nickname will only
be applied when logging into the same account it was set on.

(A custom nickname can only be set on one account. Using /nick
on a second account will overwrite the setting.)