Fix #345: (possibly) support anonymous login.
This adds an --anonymous option to /connect, which may
allow cadence to connect anonymously. Since most servers
do not allow this, it usually won't work.

(Furthermore, even if it works, the lack of a real JID will cause
the client to fail in a lot of interesting and unexpected ways.)