Fix #444 String attribute localization.
All elements' attributes can now be localized by simply giving
them the "string" class and setting the data-string-NAME attribute
to the string key that the NAME attribute should be set to.

string-html is gone.

Also, the "Clear" label is moved from settings.textColor to label.button.
to make it reusable.