Fix #94: Add emoticon search.
This fulfils a long-standing feature request by adding a search bar
to emoticon sidebars, using the Isotope JS library.

The search bar is shown in the top of the sidebar (and is sticky),
and any text entered there filters the emoticon shortcuts by name,
with a built-in latency of .3s to avoid locking down the UI.

Future feature extensions may add a tagging system for emoticons,
to allow different search terms.