Fix #421: Convert to a flexbox layout.
This rewrites the markup to get rid of absolutely positioned
page elements, and replaces it with a flexbox model.

The message pane and sidebar fill the available vertical space,
while the message pane fills the available horizontal space.
The settings and help sidebar have been remodeled to get rid of
fixed width formatting, and the settings are no longer table-based.